+++ webapp/components/general/List.js
@@ -293,12 +293,14 @@ class ListComponent extends Component {
             <Box display="flex" alignItems="center" className={classes.addIcon}>
               <IconButton
                 id={`${this.props.id}-button`}
+                disabled={this.props.disabled}
               >
                 <Add onClick={this.addChild} />
               </IconButton>
             </Box>
             <TextField
               variant="filled"
+              disabled={this.props.disabled}
               id={this.props.id}
               label={
                 this.props.label ? `Add new ${this.props.label}` : 'Add new item'
